+++ lib/maxmind/geoip2/record/traits.rb
@@ -16,9 +16,7 @@ module MaxMind
           if record && !record.key?('network') && record.key?('ip_address') &&
              record.key?('prefix_length')
             ip = IPAddr.new(record['ip_address']).mask(record['prefix_length'])
-            # We could use ip.prefix instead of record['prefix_length'], but that
-            # method only becomes available in Ruby 2.5+.
-            record['network'] = format('%s/%d', ip.to_s, record['prefix_length'])
+            record['network'] = format('%s/%d', ip.to_s, ip.prefix)
           end
         end
 
@@ -109,6 +107,28 @@ module MaxMind
         # @return [Boolean]
         def legitimate_proxy?
           get('is_legitimate_proxy')
+        end
+
+        # The {https://en.wikipedia.org/wiki/Mobile_country_code mobile country
+        # code (MCC)} associated with the IP address and ISP.
+        #
+        # This attribute is only available from the City and Insights web service
+        # and the GeoIP2 Enterprise database.
+        #
+        # @return [String, nil]
+        def mobile_country_code
+          get('mobile_country_code')
+        end
+
+        # The {https://en.wikipedia.org/wiki/Mobile_country_code mobile network
+        # code (MNC)} associated with the IP address and ISP.
+        #
+        # This attribute is only available from the City and Insights web service
+        # and the GeoIP2 Enterprise database.
+        #
+        # @return [String, nil]
+        def mobile_network_code
+          get('mobile_network_code')
         end
 
         # This is true if the IP address belongs to a public proxy. This property